ADD: Logic to setup the Ktor HttpClient
