--- data/datasource/src/main/java/android/template/api/interceptors/CatNetworkInterceptor.kt
@@ -1,11 +1,0 @@
-package android.template.api.interceptors
-
-import okhttp3.Interceptor
-
-internal fun getCatLocalInterceptor() = Interceptor { interceptorChain ->
-    val httpUrl = interceptorChain.request().url.newBuilder().apply {
-        addQueryParameter("json", "true")
-    }.build()
-    val newRequest = interceptorChain.request().newBuilder().url(httpUrl).build()
-    interceptorChain.proceed(newRequest)
-}
