+++ data/datasource/src/main/java/android/template/di/NetworkModule.kt
@@ -1,16 +1,8 @@
 package android.template.di
 
-import android.template.api.apimodels.BASE_CAT_URL
-import android.template.api.interceptors.getCatLocalInterceptor
-import android.template.api.retrofit.NetworkClient
-import android.template.api.retrofit.NetworkClientImpl
+import android.template.api.httpClient.createKtorHttpClient
 import org.koin.dsl.module
 
 val networkModule = module {
-    single<NetworkClient> {
-        NetworkClientImpl(
-            baseUrl = BASE_CAT_URL,
-            localInterceptor = getCatLocalInterceptor(),
-        )
-    }
+    single { createKtorHttpClient() }
 }
